Updated GL specification.
- Extensions added: GL_MESA_program_binary_formats, GL_MESA_tile_raster_order, GL_NV_conservative_raster_pre_snap, GLX_MESA_swap_control, GL_OES_texture_half_float, GL_QCOM_texture_foveated
- Specification fixes (GL_VERSION_1_4 requirements, gles1 API), added/updated enumerations.
- Various minor fixes.

+++ OpenGL.Net.Objects/Buffer.cs
@@ -229,7 +229,7 @@ namespace OpenGL.Objects
 				throw new InvalidOperationException("buffer is immutable");
 
 			if (Immutable && ctx.Extensions.BufferStorage_ARB) {
-				Gl.BufferStorage(Target, size, data, (uint)_UsageMask);
+				Gl.BufferStorage(Target, size, data, _UsageMask);
 				Gl.CheckErrors();
 			} else {
 				Gl.BufferData(Target, size, data, Hint);
